% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genData}
\alias{genData}
\title{Simulates data with matching variable names and characteristics}
\usage{
genData(sims, n)
}
\arguments{
\item{data}{the data frame we want to learn from.}
}
\value{
Returns a list that can be passed to genData to generate simulated data.
}
\description{
Simulates data with matching variable names and characteristics
}
\examples{
simulateData(iris)
}

