% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateData}
\alias{simulateData}
\title{Simulates data with matching variable names and characteristics}
\usage{
simulateData(data, missingness = F, discreteLength = 100)
}
\arguments{
\item{data}{the data frame we want to learn from.}
}
\value{
Returns a function taken \code{n} as an input, whcih is number of observations
}
\description{
Simulates data with matching variable names and characteristics
}
\examples{
genData <- simulateData(iris)
genData(100)
}

