% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testInteger}
\alias{testInteger}
\title{Check if everything in a vector of numbers are integers}
\usage{
testInteger(x)
}
\arguments{
\item{x}{a numbe or collection of numbers}
}
\value{
Returns \code{TRUE} if all values in vector ar integers and otherwise returns \code{FALSE}
}
\description{
Check if everything in a vector of numbers are integers
}
\examples{
testInteger(3)
testInteger(c(10, 1))
}

