% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateData}
\alias{simulateData}
\title{Simulates data with matching variable names and characteristics}
\usage{
simulateData(data, missingness = F, discreteLength = 100)
}
\arguments{
\item{data}{the data frame we want to learn from.}
}
\value{
Returns a simulated data frame.
}
\description{
Simulates data with matching variable names and characteristics
}
\examples{
simulateData(data,3)
}

